finalize new architecture
