--- services/enrichment/saver/src/plugins/README.md
@@ -1,3 +1,0 @@
-# Enrichment plugins
-
-The name of each plugin points to the information source used for enrichment procedure.
