--- services/enrichment/saver/test/unit/lib/record.json
@@ -1,47 +1,0 @@
-{
-  "computed_key": "iati/45da84c2-a349-4610-a03a-85ad86d092fd.csv.ndjson",
-  "action": "",
-  "budget": {
-    "eu_contrib": { "value": 0, "currency": "EUR", "raw": "0" },
-    "total_cost": { "value": 0, "currency": "", "raw": "" },
-    "private_fund": { "value": 0, "currency": "", "raw": "" },
-    "public_fund": { "value": 0, "currency": "", "raw": "" },
-    "other_contrib": { "value": 0, "currency": "", "raw": "" },
-    "funding_area": [],
-    "mmf_heading": ""
-  },
-  "call_year": "",
-  "description": "bar",
-  "ec_priorities": [],
-  "media": { "cover_image": "", "video": "" },
-  "programme_name": "",
-  "project_id": "2b6cdeacbf83f07fb795b393f15df20f",
-  "project_locations": [
-    {
-      "address": "",
-      "centroid": null,
-      "country_code": "BE",
-      "location": null,
-      "nuts": [],
-      "postal_code": "",
-      "region": "235",
-      "town": ""
-    }
-  ],
-  "project_website": "",
-  "related_links": [],
-  "reporting_organisation":
-    "European Commission - Development and Cooperation-EuropeAid",
-  "results": { "available": "", "result": "" },
-  "status": "",
-  "sub_programme_name": "",
-  "success_story": "",
-  "themes": [],
-  "third_parties": [],
-  "timeframe": {
-    "from": "2005-02-01T00:00:00.000Z",
-    "to": "2005-10-31T00:00:00.000Z"
-  },
-  "title": "Foo",
-  "type": []
-}
