--- services/enrichment/saver/test/unit/plugins/country/location.json
@@ -1,10 +1,0 @@
-{
-  "address": "",
-  "centroid": null,
-  "country_code": "BE",
-  "location": null,
-  "nuts": [],
-  "postal_code": "",
-  "region": "",
-  "town": ""
-}
